facevault/Backend/requirements.txt
flask
pymongo
boto3
face_recognition
opencv-python
cryptography
bcrypt
python-dotenv
